refactor(match): recover object texture coordinate ownership

+++ analysis/headers/bn_garbage_hazard_types.h
@@ -94,7 +94,7 @@ struct GarbageHazardSlot {
     GarbageHazardSlot* list_prev;
     GarbageHazardSlot* list_next;
     Vec3 bod_position;
-    int32_t render_arg_1c;
+    float render_arg_1c;
     float render_arg_20;
     void* object;
     Color4f color;
