+++ analysis/headers/path_template_types.h
@@ -44,7 +44,7 @@ typedef struct BodNode {
 typedef struct BodBase {
     BodNode bod;
     Vec3 position;
-    int32_t render_arg_1c;
+    float render_arg_1c;
     float render_arg_20;
     void* object;
     Color4f color;
@@ -58,7 +58,7 @@ typedef struct RenderableBod {
 typedef struct FringeObject {
     BodNode bod;
     Vec3 position;
-    int32_t render_arg_1c;
+    float render_arg_1c;
     float render_arg_20;
     void* object;
     Color4f color;
@@ -254,7 +254,7 @@ typedef struct VoiceManager {
 typedef struct SaltHazardSlot {
     BodNode bod;
     Vec3 bod_position;
-    int32_t render_arg_1c;
+    float render_arg_1c;
     float render_arg_20;
     void* object;
     Color4f color;
@@ -283,7 +283,7 @@ typedef struct SaltHazardSlot {
 typedef struct SubLazerSlot {
     BodNode bod;
     Vec3 bod_position;
-    int32_t render_arg_1c;
+    float render_arg_1c;
     float render_arg_20;
     void* object;
     Color4f color;
@@ -310,7 +310,7 @@ typedef struct TextureRefList {
 typedef struct TrackRowCell {
     BodNode bod;
     Vec3 anchor_position;
-    int32_t render_arg_1c;
+    float render_arg_1c;
     float render_arg_20;
     void* object;
     Color4f color;
