+++ tools/match/include/garbage_hazard_slot.h
@@ -18,7 +18,7 @@ public:
     void spawn_garbage_smoke_particle(Vector3* position, Vector3* velocity, Player* player);
 
     Vector3 bod_position;           // +0x10, BodBase::position
-    int render_arg_1c;              // +0x1c, BodBase render argument
+    float render_arg_1c;            // +0x1c, BodBase texture-u render argument
     float render_arg_20;            // +0x20, BodBase render argument
     void* object;                   // +0x24
     Color4f color;                  // +0x28
